migrated all enterprise features

+++ api/user/edit_picture.php
@@ -30,10 +30,9 @@ if(!in_array(strtolower($extension), $allowed)) {
 	]));
 }
 
-
 if(isset($_FILES["picture"]) && $_FILES["picture"]["name"]) {
 	if($file = Resources::upload("uploads", $_FILES["picture"])) {
-		if($user["picture"]) {
+		if($user["picture"] && ($user["picture"] != "user.png" && $user["picture"] != "user.jpg")) {
 			/* Delete the last pirture */
 			Resources::deleteFile("uploads/".$user["picture"]);
 		}
